requirements.txt
langgraph
langchain-openai
langchain-core
python-dotenv
fastapi
uvicorn
pandas
plotly
